[AB#33559] Move bulk publish methods to ApplicationBulkActionService
