+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application/GrantApplications/BulkApprovalsAppService.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authorization.Infrastructure;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
@@ -9,7 +8,6 @@ using System.Threading.Tasks;
 using Unity.GrantManager.Applications;
 using Unity.GrantManager.Events;
 using Unity.GrantManager.Permissions;
-using Unity.Modules.Shared;
 using Volo.Abp.EventBus.Local;
 using Volo.Abp.Uow;
 
@@ -18,8 +16,6 @@ namespace Unity.GrantManager.GrantApplications
     [Authorize]
     public class BulkApprovalsAppService(
         IApplicationRepository applicationRepository,
-        IApplicationStatusRepository applicationStatusRepository,
-        IApplicationFormRepository applicationFormRepository,
         IApplicationManager applicationManager,
         ILocalEventBus localEventBus,
         IUnitOfWorkManager unitofWorkManager) : GrantManagerAppService, IBulkApprovalsAppService
@@ -247,60 +243,6 @@ namespace Unity.GrantManager.GrantApplications
         {
             return new OperationAuthorizationRequirement { Name = $"{GrantApplicationPermissions.Applications.Default}.{triggerAction}" };
             // this should allow anyone for now, but needs to change to a specific permission
-        }
-
-        /// <summary>
-        /// Get applications for bulk publish with internal and external status information
-        /// </summary>
-        /// <param name="applicationGuids"></param>
-        /// <param name="excludePublished"></param>
-        /// <returns></returns>
-        [Authorize(UnitySelector.Application.Status.BulkPublish)]
-        public async Task<List<BulkPublishDto>> GetApplicationsForBulkPublish(Guid[] applicationGuids, bool excludePublished = true)
-        {
-            var applicationsQuery = await applicationRepository.GetQueryableAsync();
-            var statusesQuery = await applicationStatusRepository.GetQueryableAsync();
-            var formsQuery = await applicationFormRepository.GetQueryableAsync();
-
-            var results = await (
-                    from application in applicationsQuery
-                    join form in formsQuery on application.ApplicationFormId equals form.Id
-                    join status in statusesQuery on application.ApplicationStatusId equals status.Id
-                    where applicationGuids.Contains(application.Id) && (!excludePublished || !application.ExternalStatusVisibility)
-                    select new BulkPublishDto
-                    {
-                        ApplicationId = application.Id,
-                        ReferenceNo = application.ReferenceNo,
-                        ApplicantName = application.Applicant.ApplicantName ?? string.Empty,
-                        ApplicationStatus = status.InternalStatus,
-                        FormName = form.ApplicationFormName ?? string.Empty,
-                        FinalDecisionDate = application.FinalDecisionDate,
-                        ExternalStatusVisibility = application.ExternalStatusVisibility,
-                        ExternalStatus = application.ExternalStatusVisibility
-                            ? status.NotifiedStatus ?? status.ExternalStatus
-                            : status.ExternalStatus,
-                        PublishedStatus = status.NotifiedStatus ?? status.ExternalStatus,
-                    }).ToListAsync();
-
-            return results;
-        }
-
-        /// <summary>
-        /// Bulk publish applications by setting the ExternalStatusVisibility to true for the specified application GUIDs.
-        /// </summary>
-        /// <param name="applicationGuids">The GUIDs of the applications to be bulk published.</param>
-        /// <returns>A task representing the asynchronous operation.</returns>
-        [Authorize(UnitySelector.Application.Status.BulkPublish)]
-        public async Task BulkPublishApplications(Guid[] applicationGuids)
-        {
-            var applications = await applicationRepository
-                .GetListAsync(x => applicationGuids.Contains(x.Id) && !x.ExternalStatusVisibility);
-
-            foreach (var application in applications)
-            {
-                application.ExternalStatusVisibility = true;
-                await applicationRepository.UpdateAsync(application);
-            }
         }
     }
 }
